dj/requirements.txt
django-celery-beat==2.5.0
django-celery-results==2.5.0
djangorestframework==3.14.0
pika==1.3.1
pip-chill==1.0.1
redis==4.5.2
# amqp==5.1.1 # Installed as dependency for kombu
# asgiref==3.6.0 # Installed as dependency for django
# async-timeout==4.0.2 # Installed as dependency for redis
# billiard==3.6.4.0 # Installed as dependency for celery
# celery==5.2.7 # Installed as dependency for django-celery-beat, django-celery-results
# click==8.1.3 # Installed as dependency for click-plugins, click-repl, celery, click-didyoumean
# click-didyoumean==0.3.0 # Installed as dependency for celery
# click-plugins==1.1.1 # Installed as dependency for celery
# click-repl==0.2.0 # Installed as dependency for celery
# cron-descriptor==1.2.35 # Installed as dependency for django-celery-beat
# django==4.1.7 # Installed as dependency for django-celery-beat, djangorestframework, django-celery-results
# django-timezone-field==5.0 # Installed as dependency for django-celery-beat
# kombu==5.2.4 # Installed as dependency for celery
# prompt-toolkit==3.0.38 # Installed as dependency for click-repl
# python-crontab==2.7.1 # Installed as dependency for django-celery-beat
# python-dateutil==2.8.2 # Installed as dependency for python-crontab
# pytz==2022.7.1 # Installed as dependency for django-timezone-field, celery, djangorestframework
# six==1.16.0 # Installed as dependency for python-dateutil, click-repl
# sqlparse==0.4.3 # Installed as dependency for django
# tzdata==2022.7 # Installed as dependency for django-celery-beat
# vine==5.0.0 # Installed as dependency for celery, kombu, amqp
# wcwidth==0.2.6 # Installed as dependency for prompt-toolkit
